// angular-client/model/bDEEvaluateRequest.ts
/**
 * DeepBDE API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Entrada para /fragment/. - smiles: SMILES con hidrógenos explícitos - molecule_id: ID de la molécula (opcional si se proporcionan fragmentos) - bonds_idx: Lista de índices de enlaces (opcional si se proporcionan fragmentos) - export_smiles: Indica si se debe exportar la lista de SMILES de los fragmentos - export_xyz: Indica si se debe exportar el bloque XYZ de los fragmentos
 */
export interface BDEEvaluateRequest {
  smiles: string;
  molecule_id: string;
  bonds_idx: Array<number> | null;
  export_smiles?: boolean;
  export_xyz?: boolean;
}